Indices are now correct.
TODO: Have to add support for multiple generics

the way the demangler can now backtrack to properly decode symbols is
correct and does work but is very hard to read. Having to save the index
of where your Type/Path is going to sit in the `holder` then recursing
the inner path's/type's is too complex.